Moved and renamed files to more standardize naming
