+++ api/appointments/getAppointments.ts
@@ -1,7 +1,7 @@
-import handler from './handler';
-import db from './db';
+import handler from '../handler';
+import db from '../db';
 import { QueryResult } from 'pg';
-import * as auth from './authenticate';
+import * as auth from '../authenticate';
 
 export default handler(async (request: any) => {
     return await db(async (connection) => {
